aider: Fix error: Check if the "func" attribute is set before calling it, and print a usage message and exit if it is not set.
# Aider chat conversation:

USER: Fix error: Traceback (most recent call last):
  File "/Users/jzhang/dev/bilingual-tts/main.py", line 147, in <module>
    main()
  File "/Users/jzhang/dev/bilingual-tts/main.py", line 29, in main
    args.func(args)
    ^^^^^^^^^
AttributeError: 'Namespace' object has no attribute 'func'

ASSISTANT: The error message indicates that the `args` object does not have an attribute `func`. This attribute is supposed to be set by the `set_defaults` method of the argument parser for each subcommand. If the program is run without any subcommand, `args.func` will not be set, and this error will occur.

To fix this, we can add a check to see if `args.func` is set before calling it. If it is not set, we can print a usage message and exit the program.

Here is the *edit block* for this change:

```python
main.py
<<<<<<< HEAD
    # Parse the arguments
    args = parser.parse_args()

    # Call the function specified by the "func" attribute of args
    args.func(args)
=======
    # Parse the arguments
    args = parser.parse_args()

    # Check if the "func" attribute is set
    if hasattr(args, 'func'):
        # Call the function specified by the "func" attribute of args
        args.func(args)
    else:
        # Print usage message and exit
        parser.print_usage()
        exit(1)
>>>>>>> updated
```

This change will prevent
^C KeyboardInterrupt

USER: Fix error: Traceback (most recent call last):
  File "/Users/jzhang/dev/bilingual-tts/main.py", line 147, in <module>
    main()
  File "/Users/jzhang/dev/bilingual-tts/main.py", line 29, in main
    args.func(args)
    ^^^^^^^^^
AttributeError: 'Namespace' object has no attribute 'func'

ASSISTANT: The error message indicates that the `args` object does not have an attribute `func`. This attribute is supposed to be set by the `set_defaults` method of the argument parser for each subcommand. If the program is run without any subcommand, `args.func` will not be set, and this error will occur.

To fix this, we can add a check to see if `args.func` is set before calling it. If it is not set, we can print a usage message and exit the program.

Here is the *edit block* for this change:

```python
main.py
<<<<<<< HEAD
    # Parse the arguments
    args = parser.parse_args()

    # Call the function specified by the "func" attribute of args
    args.func(args)
=======
    # Parse the arguments
    args = parser.parse_args()

    # Check if the "func" attribute is set
    if hasattr(args, 'func'):
        # Call the function specified by the "func" attribute of args
        args.func(args)
    else:
        # Print usage message and exit
        parser.print_usage()
        exit(1)
>>>>>>> updated
```

This change will prevent the `AttributeError` from occurring when the program is run without any subcommand.